requirements.txt
anthropic==0.76.0
fastapi @ file:///Users/ec2-user/croot/fastapi-split_1768401105557/work
fastapi-cli @ file:///Users/ec2-user/croot/fastapi-cli_1768373425971/work
openai==1.79.0
pydantic @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pydantic_1764434463/work
pydantic-extra-types @ file:///Users/ec2-user/croot/pydantic-extra-types_1768308429644/work
pydantic-settings @ file:///opt/miniconda3/conda-bld/pydantic-settings_1764165135070/work
pydantic_core @ file:///opt/miniconda3/conda-bld/pydantic-core_1764009781890/work
uvicorn @ file:///Users/ec2-user/croot/uvicorn-split_1768226431940/work
uvloop @ file:///opt/miniconda3/conda-bld/uvloop_1762175502795/work
chromadb==1.4.1
PyPDF2 @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_e9jdhxarzu/croot/pypdf2_1733841481306/work
flashrank-0.2.10
